+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+- 0345: Asset Rendition Dispatcher framework allowing for named, cacheable renditions.
+
 ### Fixed
 - #0388: Corrected spelling of Boolean in dialog value.